feat: extract IPFS registry helper modules
